pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Yours Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
numpy = "^1.22.2"
"discord.py" = "^2.0"
pyairtable = "^1.3.0"
"python-dotenv" = "^0.21.0"
google-api-python-client = "^2.65.0"
google-auth = "^2.14.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
